Definition of done:  - update board and task module to TypeScript  - add documentation to the modules
