+++ src/resources/users/user.router.ts
@@ -6,20 +6,38 @@ import { UserProperties } from './user.model';
 
 const router = new Router({ prefix: '/users' });
 
-router.use(async (ctx, next) => {
-  ctx.type = 'application/json';
-  next();
-});
+router.use(
+  /**
+   * The function sets respose type to application/json
+   * @param ctx Koa context
+   * @param next Callbach which should be called at the end of mw function
+   */
+  async (ctx, next) => {
+    ctx.type = 'application/json';
+    next();
+  }
+);
 
-router.get('/', async (ctx) => {
-  const users = await usersService.getAll();
+router.get(
+  '/',
+  /**
+   * The function which handles get request and returns all users
+   * @param ctx Koa context
+   */
+  async (ctx) => {
+    const users = await usersService.getAll();
 
-  ctx.body = users;
-  ctx.status = 200;
-});
+    ctx.body = users;
+    ctx.status = 200;
+  }
+);
 
 router.get(
   '/:id([0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12})',
+  /**
+   * The function which handles get request and returns user by provided id
+   * @param ctx Koa context
+   */
   async (ctx) => {
     const { id } = ctx.params;
 
@@ -32,20 +50,31 @@ router.get(
       ctx.body = { message: `User wiht id ${id} is not found` };
       ctx.status = 404;
     }
-  },
+  }
 );
 
-router.post('/', async (ctx) => {
-  const { name, login, password } = ctx.request.body as UserProperties;
+router.post(
+  '/',
+  /**
+   * The function which handles post request and returns created user
+   * @param ctx
+   */
+  async (ctx) => {
+    const { name, login, password } = ctx.request.body as UserProperties;
 
-  const user = await usersService.create({ name, login, password });
+    const user = await usersService.create({ name, login, password });
 
-  ctx.body = user;
-  ctx.status = 201;
-});
+    ctx.body = user;
+    ctx.status = 201;
+  }
+);
 
 router.put(
   '/:id([0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12})',
+  /**
+   * The function which handles put request and returns updated user
+   * @param ctx Koa context
+   */
   async (ctx) => {
     const { id } = ctx.params;
     const { name, login, password } = ctx.request.body as UserProperties;
@@ -64,11 +93,15 @@ router.put(
       ctx.body = { message: `User with provided id ${id} is not found` };
       ctx.status = 404;
     }
-  },
+  }
 );
 
 router.delete(
   '/:id([0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12})',
+  /**
+   * The functions which handles delete request and delete user by id
+   * @param ctx Koa context
+   */
   async (ctx) => {
     const { id } = ctx.params;
 
@@ -81,7 +114,7 @@ router.delete(
       ctx.body = { message: `User with provided id ${id} is not found` };
       ctx.status = 404;
     }
-  },
+  }
 );
 
 export { router };
